build(cmake): allow environment to provide open62541pp
Users may wish to provide their own version of open62541pp from the
environment for a variety of versions. The obvious use case is
distributors wanting to use the dependencies from their repositories.
Therefore this commit makes CMake use a suitable version of open62541pp
if it is available, falling back to using FetchContent.